refinado e comentado ex06

+++ modulo2/ex06/asm.h
@@ -1,5 +1,5 @@
 #ifndef ASM_H
 #define ASM_H
-int swapBytes(void);
-short concatBytes(void);
+short int swapBytes(void);
+short int concatBytes(void);
 #endif
